[FIX] Remove warning on backup
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trim@66502 b456876b-0849-0410-b77d-98878d47e9d5

+++ src/libs/trim/Backup.php
@@ -142,7 +142,7 @@ class Backup
         $backupDir = $backupDir ?: $this->backupDir;
         $sqlpath = "{$backupDir}/database_dump.sql";
 
-        unlink($sqlpath);
+        file_exists($sqlpath) && unlink($sqlpath);
         $app->backupDatabase($sqlpath);
 
         if (file_exists($sqlpath)) {
